refactor(server): drop UpgradeCursorService, fold cursor into the metadata mutator
UpgradeSequenceRunnerService already advances a local `cursor` variable
during run() and the DB row in core.upgradeMigration is the source of
truth. UpgradeCursorService was a parallel in-memory layer that
duplicated both. UpgradeAwareEntityMetadataService now owns its cursor
field directly, exposes refresh() that reads the truth from
UpgradeMigrationService.getLastAttemptedInstanceCommand, and is called
by the runner before/after each instance step (and in the run's
finally). Boot-time decorator validation also moves into the mutator
since it owns the sequence-step lookup map.
